fixed type isues

+++ types/types/number.d.ts
@@ -1,3 +1,3 @@
-declare const toNumber: typeof parseFloat;
+declare const toNumber: typeof parseInt;
 declare const isNumber: (val: any) => boolean;
 export { toNumber, isNumber };
